+++ doc/release-notes.md
@@ -3,3 +3,4 @@ Bitcoin ABC version 0.17.3 is now available from:
   <https://download.bitcoinabc.org/0.17.3/>
 
 This release includes the following features and fixes:
+ - Remove the bip9params configuration.